Add files via upload
A glove integrated with flex sensors and an MPU6050 accelerometer that converts ASL gestures to speech in real-time for specially-abled people.  Made with the help of Arduino microcontroller that converts Analog input to digital input. Python is embedded with the IDE with pyserial to communicate and translate the sign language.